refactor: timetable service routting code separate
